Allows user login attribute to be configured
This change abstracts the primary user login attribute / identifier to a configurable option (that defaults to "samaccountname"), instead of the previously hard-coded options of either GUID, samaccountname, or userPrincipalName. In many places (like the find method) this was simply set as samaccountname, which might not be a searchable attribute in all organizations.