Fix auth with invalid credentials exception
Issue: User authentication seems to work fine if the credentials are correct. However, when invalid credentials are presented to the Adldap->authenticate(), an Exception is thrown. As a side-effect, the rebind is never done, so subsequent auth attempts always fail. An Exception is probably fine for invalid administrator bind credentials, but if a user mis-types their password, it should probably just return false.

Fix: Catch the exception and return false. Also, add the check to see if administrator credentials are null before rebinding (to support anonymous bind).